Rename: constants.js Mypage의 최상위폴더로 이동
